server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
transformers==4.46.0
torch==2.5.0
huggingface_hub==0.26.0
requests==2.32.0
python-dotenv==1.0.1
accelerate==1.0.0
